requirements/cpu_requirements.txt
numpy>=1.17.0
numba
scipy>=1.6.0
pyyaml>=5.3
inflect==5.3.0
Unidecode==1.3.2
mindspore-dev
tqdm
wget
sentencepiece
easydict
matplotlib
Levenshtein
pylint
pytest
